+++ src/pystormtracker/__init__.py
@@ -11,7 +11,7 @@ from .simple.tracker import SimpleTracker
 try:
     __version__ = version("pystormtracker")
 except PackageNotFoundError:
-    __version__ = "0.7.0.dev0"
+    __version__ = "unknown"
 
 
 __all__ = [